Cleanup URI tests
